Refaktorering av typer. Flyttet begrunnelse og resultat opp til behandling.
