+++ dev-server/mock/fagsak-1.json
@@ -12,6 +12,8 @@
                 "status": "OPPRETTET",
                 "kategori": "NASJONAL",
                 "opprettetTidspunkt": "2020-02-01T12:00:00.00",
+                "resultat": "INNVILGET",
+                "begrunnelse": "Mock begrunnelse",
                 "personer": [
                     {
                         "type": "SØKER",
@@ -88,9 +90,7 @@
                                 "beløp": 1054,
                                 "stønadFom": "2020-01-01"
                             }
-                        ],
-                        "resultat": "INNVILGET",
-                        "begrunnelse": "Fake information"
+                        ]
                     }
                 ]
             },
@@ -101,6 +101,8 @@
                 "status": "FERDIGSTILT",
                 "kategori": "NASJONAL",
                 "opprettetTidspunkt": "2020-02-01T12:00:00.00",
+                "resultat": "INNVILGET",
+                "begrunnelse": "Mock begrunnelse",
                 "personer": [
                     {
                         "type": "SØKER",
@@ -152,9 +154,7 @@
                                 "beløp": 1054,
                                 "stønadFom": "2020-01-01"
                             }
-                        ],
-                        "resultat": "INNVILGET",
-                        "begrunnelse": "Fake information"
+                        ]
                     }
                 ]
             }
