+++ src/frontend/komponenter/Fagsak/Vedtak/OppsummeringVedtak.tsx
@@ -5,10 +5,11 @@ import * as React from 'react';
 import { useHistory } from 'react-router';
 import { axiosRequest } from '../../../api/axios';
 import { hentAktivVedtaksbrev } from '../../../api/oppsummeringvedtak';
-import { IFagsak, IBehandling, BehandlingStatus } from '../../../typer/fagsak';
+import { IFagsak, BehandlingStatus } from '../../../typer/fagsak';
 import { Ressurs, RessursStatus } from '../../../typer/ressurs';
 import { AxiosError } from 'axios';
 import { useFagsakDispatch, actions } from '../../FagsakProvider';
+import { IBehandling } from '../../../typer/behandling';
 
 interface IVedtakProps {
     fagsak: IFagsak;
@@ -23,9 +24,9 @@ const OppsummeringVedtak: React.FunctionComponent<IVedtakProps> = ({ fagsak }) =
 
     const [submitFeil, settSubmitFeil] = React.useState('');
 
-    const aktivBehandling = fagsak.behandlinger.find(
-        (behandling: IBehandling) => behandling.aktiv === true
-    );
+    const aktivBehandling = fagsak.behandlinger.find((behandling: IBehandling) => {
+        return behandling.aktiv === true;
+    });
 
     React.useEffect(() => {
         if (
