+++ src/frontend/komponenter/Fagsak/fagsak.less
@@ -2,6 +2,6 @@
     &__content {
         width: 60rem;
         margin: 0 auto;
-        padding-top: 2rem;
+        padding: 2rem 0;
     }
 }
