+++ src/frontend/komponenter/Felleskomponenter/Informasjonsbolk/Informasjonsbolk.tsx
@@ -14,7 +14,7 @@ const Informasjonsbolk: React.FunctionComponent<IProps> = ({ informasjon }) => {
     return (
         <div className={'informasjonsbolk'}>
             {informasjon.map((info: IInformasjon) => {
-                return <Element key={info.label} children={info.label} />;
+                return <Element key={info.label + info.tekst} children={info.label} />;
             })}
             {informasjon.map((info: IInformasjon) => {
                 return <Normaltekst key={info.tekst} children={info.tekst} />;
