+++ src/frontend/komponenter/app.less
@@ -1,5 +1,5 @@
 .container {
-    height: calc(100% - 4rem);
+    height: calc(100% - 2rem);
     overflow-y: auto;
     overflow-x: hidden;
 }
