+++ src/frontend/typer/common.ts
@@ -1,0 +1,6 @@
+export interface INøkkelPar {
+    [key: string]: {
+        id: string;
+        navn: string;
+    };
+}
