+++ src/frontend/typer/person.ts
@@ -37,37 +37,3 @@ export enum PersonType {
     ANNENPART = 'ANNENPART',
     BARN = 'BARN',
 }
-
-export interface IPersonhistorikk {
-    adresser: IPersonAdresse[];
-    statsborgerskap: IPersonStatsborgerskap[];
-}
-
-export interface IPersonAdresse {
-    adresseType: AdresseType;
-    adresselinje1?: string;
-    adresselinje2?: string;
-    adresselinje3?: string;
-    adresselinje4?: string;
-    land: string;
-    periode: IPeriode;
-    postnummer?: string;
-    poststed?: string;
-}
-
-export interface IPersonStatsborgerskap {
-    periode: IPeriode;
-    statsborgerskap: string;
-}
-
-export interface IPersonRelasjon {
-    fraPersonIdent: string;
-    harSammeBosted: boolean;
-    relasjonsrolle: RelasjonsRolleType;
-    tilPersonIdent: string;
-}
-
-export interface IPeriode {
-    fomDato: string;
-    tomDato: string;
-}
