+++ src/frontend/typer/vedtak.ts
@@ -1,0 +1,8 @@
+import { IBarnBeregning } from './behandle';
+
+export interface IVedtakForBehandling {
+    aktiv: boolean;
+    ansvarligSaksbehandler: string;
+    barnasBeregning: IBarnBeregning[];
+    vedtaksdato: string;
+}
